Rework everything to support es6 modules as build target
The goal of this changeset is to allow simpler and more modern inclusion patterns.

Best of all it should now be possible to do import {filter, map} from 'iter-tools';
and let your packager take care of tree shaking away the rest of the lib.

I have removed npm-release, as it seems that the npm version command is meant to fulfill
all the same major functions.